Avoid generated dependency in credential test
